file loader upd

+++ global/parts/user_card/user_card.js
@@ -32,10 +32,12 @@ function process_user_card() {
 
     const deauth_button = document.createElement('button')
     button_row.appendChild(deauth_button)
-    deauth_button.innerText = '⇐ Вийти'
+    deauth_button.innerText = 'Вийти'
     deauth_button.addEventListener('click', () => {
-        sessionStorage.clear()
-        window.location.href = '/';
+        if (confirm('Ви точно бажаєте вийти зі свого профілю?')) {
+            sessionStorage.clear()
+            window.location.href = '/';
+        }
     })
 
     const info_button = document.createElement('button')
@@ -48,39 +50,39 @@ function process_user_card() {
         document.body.appendChild(overlay)
         const shade = document.createElement('div')
         overlay.appendChild(shade)
-        shade.className='shade'
+        shade.className = 'shade'
         const info_block = document.createElement('div')
         info_block.className = 'info_block'
         overlay.appendChild(info_block)
-        shade.addEventListener('click',()=>{
-            overlay.style.animation='fadeOut 0.3s'
-            overlay.addEventListener('animationend',()=>{
+        shade.addEventListener('click', () => {
+            overlay.style.animation = 'fadeOut 0.3s'
+            overlay.addEventListener('animationend', () => {
                 overlay.remove()
             })
         })
-        const fields=[
-            {name:`Ім'я:`,value:''},
-            {name:`Логін:`,value:account.login},
-            {name:`Ранг:`,value:acc_level[account.permission_level]},
-            {name:`Секретний ключ:`,value:account.skey}
+        const fields = [
+            { name: `Ім'я:`, value: '' },
+            { name: `Логін:`, value: account.login },
+            { name: `Ранг:`, value: acc_level[account.permission_level] },
+            { name: `Секретний ключ:`, value: account.skey }
         ]
         if (!account.name && !account.surname) {
             fields[0].value = 'Відсутнє'
         } else {
             fields[0].value = usr_name
         }
-        fields.forEach(field=>{
-            const fblock=document.createElement('div')
+        fields.forEach(field => {
+            const fblock = document.createElement('div')
             info_block.appendChild(fblock)
-            fblock.className='fblock'
-            const fname=document.createElement('div')
+            fblock.className = 'fblock'
+            const fname = document.createElement('div')
             fblock.appendChild(fname)
-            fname.className='fname'
-            fname.innerText=field.name
-            const fval=document.createElement('div')
+            fname.className = 'fname'
+            fname.innerText = field.name
+            const fval = document.createElement('div')
             fblock.appendChild(fval)
-            fval.className='fval'
-            fval.innerText=field.value
+            fval.className = 'fval'
+            fval.innerText = field.value
         })
     })
